+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `deviceId` is now stored with the api
+
+### Changed
+
+- `session` now throws when the token isn't a reMarkable jwt
+
 ## [14.0.0] - 2026-08-17
 
 ### Changed